relocate static files to assets folder, delete unnecessary static data, and fix minor css issue in home page
