--- client/src/assets/images/index.js
@@ -1,3 +1,0 @@
-import { CloseSVG } from "./close";
-import { SearchSVG } from "./search";
-export { CloseSVG, SearchSVG };
