+++ client/src/components/DropDownSelect.js
@@ -15,7 +15,7 @@ function DropDownSelect({ items, label, chooseFilterCallback }) {
     };
 
     return (
-        <FormControl sx={{ m: 1, minWidth: 200, mt: 1}}>
+        <FormControl sx={{ m: 1, minWidth: 200, my: 1}}>
             <InputLabel id={label + "-select-label"}>{label}</InputLabel>
             <Select
                 id={label + "-select"}
